GenomeDescriptions class becoems its own module
